+++ docs/release-notes/change-log.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.1.17
+
+### Fixed
+* Missing decrement on metrics `substreams_active_requests`
+
+## v1.1.16
 
 ### Added
+* `substreams_active_requests` and `substreams_counter` metrics to `substreams-tier1`
 
 ### Changed